scripts/docker-main/requirements.txt
bitstring==4.1.4
boto3==1.34.149
psycopg2==2.9.9
python_logging_loki==0.3.1
Requests==2.32.3
